+++ src/ts/wysiwyg/processKeydown.ts
@@ -64,48 +64,65 @@ export const processKeydown = (vditor: IVditor, event: KeyboardEvent) => {
 
     // md 处理
     const pElement = hasClosestByMatchTag(startContainer, "P");
-    if (pElement && !isCtrl(event) && !event.altKey && event.key === "Enter") {
-        const pText = String.raw`${pElement.textContent}`.replace(/\\\|/g, "").trim();
-        const pTextList = pText.split("|");
-        if (pText.startsWith("|") && pText.endsWith("|") && pTextList.length > 3) {
-            // table 自动完成
-            let tableHeaderMD = pTextList.map(() => "---").join("|");
-            tableHeaderMD = pElement.textContent + tableHeaderMD.substring(3, tableHeaderMD.length - 3) + "\n|<wbr>";
-            pElement.outerHTML = vditor.lute.SpinVditorDOM(tableHeaderMD);
-            setRangeByWbr(vditor.wysiwyg.element, range);
-            afterRenderEvent(vditor);
-            scrollCenter(vditor.wysiwyg.element);
-            event.preventDefault();
-            return true;
-        }
-
-        // hr 渲染
-        if (isHrMD(pElement.innerHTML)) {
-            // 软换行后 hr 前有内容
-            let pInnerHTML = "";
-            const innerHTMLList = pElement.innerHTML.trimRight().split("\n");
-            if (innerHTMLList.length > 1) {
-                innerHTMLList.pop();
-                pInnerHTML = `<p data-block="0">${innerHTMLList.join("\n")}</p>`;
+    if (pElement) {
+        if (!isCtrl(event) && !event.altKey && event.key === "Enter") {
+            const pText = String.raw`${pElement.textContent}`.replace(/\\\|/g, "").trim();
+            const pTextList = pText.split("|");
+            if (pText.startsWith("|") && pText.endsWith("|") && pTextList.length > 3) {
+                // table 自动完成
+                let tableHeaderMD = pTextList.map(() => "---").join("|");
+                tableHeaderMD =
+                    pElement.textContent + tableHeaderMD.substring(3, tableHeaderMD.length - 3) + "\n|<wbr>";
+                pElement.outerHTML = vditor.lute.SpinVditorDOM(tableHeaderMD);
+                setRangeByWbr(vditor.wysiwyg.element, range);
+                afterRenderEvent(vditor);
+                scrollCenter(vditor.wysiwyg.element);
+                event.preventDefault();
+                return true;
             }
 
-            pElement.insertAdjacentHTML("afterend", pInnerHTML + '<hr data-block="0"><p data-block="0">\n<wbr></p>');
-            pElement.remove();
-            setRangeByWbr(vditor.wysiwyg.element, range);
-            afterRenderEvent(vditor);
-            scrollCenter(vditor.wysiwyg.element);
-            event.preventDefault();
-            return true;
+            // hr 渲染
+            if (isHrMD(pElement.innerHTML)) {
+                // 软换行后 hr 前有内容
+                let pInnerHTML = "";
+                const innerHTMLList = pElement.innerHTML.trimRight().split("\n");
+                if (innerHTMLList.length > 1) {
+                    innerHTMLList.pop();
+                    pInnerHTML = `<p data-block="0">${innerHTMLList.join("\n")}</p>`;
+                }
+
+                pElement.insertAdjacentHTML("afterend",
+                    `${pInnerHTML}<hr data-block="0"><p data-block="0">\n<wbr></p>`);
+                pElement.remove();
+                setRangeByWbr(vditor.wysiwyg.element, range);
+                afterRenderEvent(vditor);
+                scrollCenter(vditor.wysiwyg.element);
+                event.preventDefault();
+                return true;
+            }
+
+            if (isHeadingMD(pElement.innerHTML)) {
+                // heading 渲染
+                pElement.outerHTML = vditor.lute.SpinVditorDOM(pElement.innerHTML + '<p data-block="0">\n<wbr></p>');
+                setRangeByWbr(vditor.wysiwyg.element, range);
+                afterRenderEvent(vditor);
+                scrollCenter(vditor.wysiwyg.element);
+                event.preventDefault();
+                return true;
+            }
         }
 
-        if (isHeadingMD(pElement.innerHTML)) {
-            // heading 渲染
-            pElement.outerHTML = vditor.lute.SpinVditorDOM(pElement.innerHTML + '<p data-block="0">\n<wbr></p>');
+        // 软换行会被切割 https://github.com/Vanessa219/vditor/issues/220
+        if (pElement.previousElementSibling && event.key === "Backspace" && !isCtrl(event) && !event.altKey &&
+            !event.shiftKey && pElement.textContent.trimRight().split("\n").length > 1 &&
+            getSelectPosition(pElement, range).start === 0) {
+            const lastElement = getLastNode(pElement.previousElementSibling) as HTMLElement
+            if (!lastElement.textContent.endsWith('\n')) {
+                lastElement.textContent = lastElement.textContent + '\n'
+            }
+            lastElement.parentElement.insertAdjacentHTML("beforeend", `<wbr>${pElement.innerHTML}`);
+            pElement.remove();
             setRangeByWbr(vditor.wysiwyg.element, range);
-            afterRenderEvent(vditor);
-            scrollCenter(vditor.wysiwyg.element);
-            event.preventDefault();
-            return true;
         }
     }