+++ CHANGELOG.md
@@ -49,11 +49,11 @@
 
 * [open issues](https://github.com/Vanessa219/vditor/issues)
 
-### v3.0.2 / 未发布
+### v3.0.3 / 未发布
 
 * [27](https://github.com/Vanessa219/vditor/issues/27) 支持类似 Typora 的及时渲染模式 `引入特性`
 
-### v3.0.1 / 2020-03-19
+### v3.0.2 / 2020-03-19
 
 * [225](https://github.com/Vanessa219/vditor/pull/225) publish type declaration file `引入特性`
 * [224](https://github.com/Vanessa219/vditor/issues/224) md2html 方法报错 `修复缺陷`